exp: Fix filters on aggregation mode
