+++ ui/src/plugins/dev.perfetto.ExplorePage/query_builder/nodes/aggregation_node.ts
@@ -26,7 +26,8 @@ import {
   columnInfoFromName,
   newColumnInfoList,
 } from '../column_info';
-import {createFiltersProto} from '../operations/operation_component';
+import {createFiltersProto, FilterOperation} from '../operations/filter';
+import {FilterDefinition} from '../../../../components/widgets/data_grid/common';
 import {MultiselectInput} from '../../../../widgets/multiselect_input';
 import {Select} from '../../../../widgets/select';
 import {TextInput} from '../../../../widgets/text_input';
@@ -144,11 +145,22 @@ export class AggregationNode implements QueryNode {
   }
 
   nodeSpecificModify(): m.Child {
-    return m(AggregationOperationComponent, {
-      groupByColumns: this.state.groupByColumns,
-      aggregations: this.state.aggregations,
-      onchange: this.state.onchange,
-    });
+    return m(
+      '.node-specific-modify',
+      m(AggregationOperationComponent, {
+        groupByColumns: this.state.groupByColumns,
+        aggregations: this.state.aggregations,
+        onchange: this.state.onchange,
+      }),
+      m(FilterOperation, {
+        filters: this.state.filters,
+        sourceCols: this.finalCols,
+        onFiltersChanged: (newFilters: ReadonlyArray<FilterDefinition>) => {
+          this.state.filters = newFilters as FilterDefinition[];
+          this.state.onchange?.();
+        },
+      }),
+    );
   }
 
   clone(): QueryNode {
@@ -175,41 +187,36 @@ export class AggregationNode implements QueryNode {
       this.state.groupByColumns,
       this.state.aggregations,
     );
-    const filtersProto = createFiltersProto(
-      this.state.filters,
-      this.sourceCols,
-    );
+    const filtersProto = createFiltersProto(this.state.filters, this.finalCols);
 
     // If the previous node already has an aggregation, we need to create a
     // subquery.
+    let sq: protos.PerfettoSqlStructuredQuery;
     if (prevSq.groupBy) {
-      const sq = new protos.PerfettoSqlStructuredQuery();
-      sq.id = this.nodeId;
+      sq = new protos.PerfettoSqlStructuredQuery();
+      sq.id = nextNodeId();
       sq.innerQuery = prevSq;
-      if (filtersProto) sq.filters = filtersProto;
-      if (groupByProto) sq.groupBy = groupByProto;
-      const selectedColumns = createSelectColumnsProto(this);
-      if (selectedColumns) sq.selectColumns = selectedColumns;
-      return sq;
+    } else {
+      sq = prevSq;
     }
 
-    // Otherwise, we can just add the aggregation and filters to the previous
-    // query.
-    if (filtersProto) {
-      if (prevSq.filters !== undefined) {
-        prevSq.filters.push(...filtersProto);
-      } else {
-        prevSq.filters = filtersProto;
-      }
-    }
     if (groupByProto) {
-      prevSq.groupBy = groupByProto;
+      sq.groupBy = groupByProto;
     }
     const selectedColumns = createSelectColumnsProto(this);
     if (selectedColumns) {
-      prevSq.selectColumns = selectedColumns;
+      sq.selectColumns = selectedColumns;
     }
-    return prevSq;
+
+    if (filtersProto) {
+      const outerSq = new protos.PerfettoSqlStructuredQuery();
+      outerSq.id = this.nodeId;
+      outerSq.innerQuery = sq;
+      outerSq.filters = filtersProto;
+      return outerSq;
+    }
+
+    return sq;
   }
 }
